add check if `function` folder exist? then delete
Added a check to createCloudFunction.js to see if the `function` folder
alread exists before trying to create it. If it does exist, then the
folder is deleted. This prevents an error when using `mkdir` to create a
folder with the same name as a folder that already exists.

Also added a few console logs to report when the folder is being delted
and created.

Co-authored-by: Mat Sachs <matochondrion@gmail.com>
Co-authored-by: Linus Phan <phanlinus@gmail.com>
Co-authored-by: Jacob Coker <jacobleecd@gmail.com>